refactor: minor formatting, remove misleading ヶヵtoRomaji conversions
The ヶヵforms can stand in for "ka, ga, or ko" and are effectively being used like symbols/kanji (and not as kana) when seen in 一ヶ月 etc.
Therefore we should make no assumptions about converting them. However, allowing the individual small forms to be entered specifically via (x|l)(ke|ka|wa) in toKana() is still a valid use case and the functionality has been retained.