replace or remove added attributes when element is unbound
when the input is bound attributes are added or overwritten. When the
input is unbound this commit reinstates overridden attributes and
removes added attributes.

I opted to store all attributes that were added or manipulated in a map
and save the map in a data attribute that could be referenced when the
element is unbound. I felt this approach was easier than only adding
overridden attributes to the map and having some conditional logic to
determine whether to remove or replace.  This way the unbinding code
shouldn't get out of step with the binding code as any attribute added
in bind will get the correct treatment when unbound.

I ran in to an issue with cypress where i wanted to write e.g.
`.should('not.have.attr', 'lang')`, but that seemed to break the
chaining as the subject was replaced.