build(package.json, rollup.config, and release.js): added typescript compiler for type definitions
As part of adding Typescript, a number of dependencies had to be updated. All @babel packages and
@rollup packages were updated. Additionally prettier-eslint-cli had to be updated due to a typing
issue in the dependency tree. This also meant that eslint also had to be updated to maintain
compatibility. Due to these package changes the minumum node devEngines version has been bumped to
16. A tsconfig.json file has been added, as well as renaming the base index file from index.js to
index.ts.

Adds packages to enable #141