refactor(function type definitions): function typings updated for type definition output
constants.js has been changed to a typescript file to enable the exporting of the DefaultOptions
type. This is then used by various function definitions, which provides type checking for the
parameters. Regexp has been changed to RegExp as Regexp was resolving to the any type. Other
parameters defined as Object have been specified with types to match their expected input.
TOKEN_TYPES's export was removed as it was unused and was being generated into the d.ts file.

Resolve external typings for #141