build(jsdocs): modified JSDoc generation to support new typescript files and typings
Added better-docs to piggyback off of its typescript plugin. Updated instances of string to String
and boolean to Boolean. This matches the rest of the existing documentation. Typescript types are
unaffected by these changes and automatically get converted to the correct typescript type. Changed
instances of [index: string]: string to Object.<string, string> JSDoc type. This gets converted in
typescript, but the previous syntax was not supported in JSDocs.

#141